refactor(admin): Move the neolace-admin client to its own root folder
