+++ backend/neolace/deps/neolace-sdk.ts
@@ -1,2 +1,1 @@
 export * from "../../../neolace-sdk/src/index.ts";
-export { getApiClientFromEnv } from "../../../neolace-sdk/src/cli-client.ts";
