+++ neolace-admin/deno.jsonc
@@ -1,0 +1,16 @@
+{
+    "lint": {
+        "rules": {
+            "exclude": [
+                "require-await"
+            ]
+        }
+    },
+    "fmt": {
+        "options": {
+            "indentWidth": 4,
+            "lineWidth": 120
+        }
+    },
+    "lock": false
+}
